+++ product/agent/collector/src/integrationTest/java/se/crisp/codekvast/agent/collector/itest/CollectorIntegrationTest.java
@@ -1,0 +1,18 @@
+package se.crisp.codekvast.agent.collector.itest;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.runners.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class CollectorIntegrationTest {
+
+    @Test
+    public void should_have_registered_method_executions() throws Exception {
+        // given
+
+        // when
+
+        // then
+    }
+}